Add support for custom ingress CA chain in OpenShift
In OpenShift, it is possible to configure a custom certificate for the
ingress routes. The full CA chain is provided to allow certificate
verification. By default, the custom CA chain is not exposed in
`/var/run/secrets/kubernetes.io/serviceaccount/ca.crt`, so pods cannot
reach services on their external URL.

The UI uses a callback link with the external URL of the UI, since it
will then be passed to the user's browser that can only access external
URLs. This forbids the access to the UI when using custom CA.

This change applies the OpenShift documentation [1] to mount the custom CA
certificate merged with the other internal CA certificates in the UI
pod. This is done by creating an empty ConfigMap annotated with
`config.openshift.io/inject-trusted-cabundle="true"` and mounting this
ConfigMap in the UI pod.

We also need to update the forklift-ui entrypoint to merge the service
service CA certificate with the merged CA bundle into
`/opt/app-root/src/ca.crt`.

[1] https://docs.openshift.com/container-platform/4.9/networking/configuring-a-custom-pki.html

Signed-off-by: Fabien Dupont <fabiendupont@pm.me>